[JENKINS-73904] Increase test codecoverage and fix issues
